# docs/swagger.yaml
basePath: /
definitions:
...
  result.Result-string:
    properties:
      code:
        type: integer
      data:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  router.WebDavLoginJson:
    properties:
      host:
        type: string
      passwd:
        type: string
      user:
        type: string
    type: object
host: 127.127.127.127:12777
info:
  contact:
    name: API Support
    url: https://github.com/miru-project
  description: Miru Core API documentation
  license:
    name: MIT
    url: https://github.com/miru-project/miru-core/blob/main/LICENSE
  title: Miru Core API
  version: "1.0"
paths:
...
  /drive/backup:
    get:
      description: Backup the Miru database to a connected WebDav server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result-string'
        "500":
          description: Backup failed
          schema:
            $ref: '#/definitions/result.Result-string'
      summary: Backup database to WebDav
      tags:
      - webdav
  /drive/login:
    post:
      consumes:
      - application/json
      description: Authenticate with a WebDav server
      parameters:
      - description: WebDav login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/router.WebDavLoginJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result-string'
        "400":
          description: Invalid JSON or missing required fields
          schema:
            $ref: '#/definitions/result.Result-string'
      summary: Login to WebDav server
      tags:
      - webdav
  /drive/restore:
    get:
      description: Restore the Miru database from a backup on a connected WebDav server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result-string'
        "500":
          description: Restore failed
          schema:
            $ref: '#/definitions/result.Result-string'
      summary: Restore database from WebDav
      tags:
      - webdav
...
schemes:
- http
- https
swagger: "2.0"